+++ include/eEVM/globalstate.h
