+++ include/eEVM/simple/simplestorage.h
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "../include/storage.h"
+#include "eEVM/storage.h"
 
 #include <map>
