+++ src/account.cpp
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "../include/account.h"
+#include "eEVM/account.h"
 
-#include "../include/exception.h"
+#include "eEVM/exception.h"
 
 #include <ostream>
 #include <type_traits>
