+++ src/transaction.cpp
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "../include/transaction.h"
+#include "eEVM/transaction.h"
 
-#include "../include/util.h"
+#include "eEVM/util.h"
 
 namespace eevm
 {
