+++ tests/harness.cpp
@@ -1,13 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "../include/disassembler.h"
-#include "../include/opcode.h"
-#include "../include/processor.h"
-#include "../include/rlp.h"
-#include "../include/simple/simpleglobalstate.h"
-#include "../include/simple/simplestorage.h"
-#include "../include/util.h"
+#include "eEVM/disassembler.h"
+#include "eEVM/opcode.h"
+#include "eEVM/processor.h"
+#include "eEVM/rlp.h"
+#include "eEVM/simple/simpleglobalstate.h"
+#include "eEVM/simple/simplestorage.h"
+#include "eEVM/util.h"
 
 #include <doctest/doctest.h>
 #include <fstream>
