+++ tests/rlp.cpp
@@ -1,6 +1,9 @@
-#include "../include/rlp.h"
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
 
-#include "../include/util.h"
+#include "eEVM/rlp.h"
+
+#include "eEVM/util.h"
 
 #include <doctest/doctest.h>
